requirements/dev.txt
# =============================================================================
# Development Requirements
# Includes base requirements plus development tools
# =============================================================================

-r base.txt

# =============================================================================
# Debug Tools
# =============================================================================
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=20.0.0
coverage>=7.3.0

# =============================================================================
# Code Quality
# =============================================================================
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
pylint>=3.0.0
mypy>=1.7.0
django-stubs>=4.2.0

# =============================================================================
# Development Server
# =============================================================================
django-browser-reload>=1.12.0

# =============================================================================
# Documentation
# =============================================================================
Sphinx>=7.2.0

# =============================================================================
# Jupyter (for data exploration)
# =============================================================================
ipython>=8.17.0
